Fix a compiler warning in MapCommand.cs

+++ OpenRA.Mods.Common/UtilityCommands/MapCommand.cs
@@ -76,13 +76,10 @@ namespace OpenRA.Mods.Common.UtilityCommands
 									map.Save(new Folder(z.Name.Replace(".oramap", "")));
 								break;
 							case "repack":
-								if (mapPackage is Folder f)
+								if (mapPackage is Folder f && File.Exists(f.Name + ".oramap"))
 								{
-									if (File.Exists(f.Name + ".oramap"))
-									{
-										map.Save(ZipFileLoader.Create(f.Name + ".oramap"));
-										Directory.Delete(f.Name, true);
-									}
+									map.Save(ZipFileLoader.Create(f.Name + ".oramap"));
+									Directory.Delete(f.Name, true);
 								}
 
 								break;
